+++ tests/e2e/cases/process_stdio_write_ignored_arguments/in.ts
@@ -15,4 +15,5 @@ const ok = process.stdout.write("global:", "utf8", done, mark("g"));
 stdout.write("", "utf8", done, mark("n"));
 processModule.stdout.write("", "utf8", done, mark("m"));
 proc.stdout.write("", "utf8", done, mark("d"));
+stdout.write("", done, mark("s"));
 console.log("write ignored:", ok, seen);